Fixed issue #86. Projects load only when necessary
There was a bug in jquery (http://bugs.jquery.com/ticket/13976) preventing the
insertion of rendered HTML (from a template) into the DOM, so I switched to
using jquery 1.11.3. If this is a problem, the original jquery link is still
in header.html, commented out.

Also, the HiSim tool randomly broke...will fix this next. Might be due to
additions to the result table, which have also broken view old results for
HiSim.