Fix a bug where null results could be viewed
This should resolve issue #87 and some of the seg faults caused by viewing
results that are no longer present in the results directory.
There may still be a case where null HiSim results can be viewed since even
when the tool fails, it still writes some output to the results directory.