Readd search size filter, client+server side code
Looks like lots of the javascript slider code conflicts with one another. Bootstrap slider and the jquery-ui slider fight for style+animation. Had to remove bootstrap slider links from the header and add them to the set threshold page so my changes didn't break that page.

The search slider now uses the jquery-ui slider and can properly be styled to look like the rest of the search page.