Alter the geneset annotation behavior on edit geneset page
The tree view now only allows users to add new annotations of the types
"Manual Association" or "Manual Rejection" to the geneset. Previous behavior
allowed users to remove annotations or add new ones of whatever reference type
was selected. "Manual Rejection" will be used in cases where a user has
positively deselected an annotation and it should not be reselected (e.g. in
the case of running the automated annotator again).

Added two new functions to geneweaverdb for updating a current annotation
reference type and another for determining if a particular term is annotated to
a geneset.